[MCS-Rich Media] : Create File View and handling

+++ netomichatsdk/src/main/java/com/netomi/chat/model/theme/TypingIndicator.kt
@@ -2,6 +2,6 @@ package com.netomi.chat.model.theme
 
 data class TypingIndicator(
     val enabled: Boolean,
-    val maxTime: Int,
-    val minTime: Int
+    val maxTime: Long,
+    val minTime: Long
 )
